requirements.txt
langchain
langchain-community
langchain-together
langchain-core
langchain-huggingface
langsmith
torch
openai
PyMuPDF
pdf2image
mammoth
python-docx
surya-ocr
FlagEmbedding
streamlit
transformers
sentence-transformers
tqdm
underthesea
qdrant-client
python-dotenv
cohere
